+++ lib/forEach.d.ts
@@ -3,21 +3,21 @@ import { GenericObject } from "./Interfaces";
 export interface ForEachCallbackType extends Function {
     /**
      * Is called for every element of the iterated object
-     * @param value The current value
-     * @param key The current key
-     * @param iteratedObject The iterated object
-     */
-    (value?: any, key?: string | number, iteratedObject?: any): void | boolean;
-}
-export interface ForEachCallbackType extends Function {
-    /**
-     * Is called for every element of the iterated object
      * @param $value The current value as a jquery object
      * @param value The current value
      * @param key The current key
      * @param iteratedObject The iterated object
      */
-    ($value?: JQuery, key?: any, value?: any, iteratedObject?: any): void | boolean;
+    ($value?: JQuery | any, key?: any, value?: any, iteratedObject?: any): void | boolean;
+}
+export interface ForEachCallbackType extends Function {
+    /**
+     * Is called for every element of the iterated object
+     * @param value The current value
+     * @param key The current key
+     * @param iteratedObject The iterated object
+     */
+    (value?: any, key?: string | number, iteratedObject?: any): void | boolean;
 }
 /**
  * Loops over arrays or objects and applies a given callback